+++ sphae.sh
@@ -6,15 +6,19 @@
 #SBATCH --error=%x-%j.err.txt
 #SBATCH --time=1-0
 #SBATCH --ntasks=1
-#SBATCH --cpus-per-task=64
-#SBATCH --mem=500G
-#SBATCH --partition=high-capacity
-#SBATCH --qos=hc-concurrent-jobs
+#SBATCH --cpus-per-task=32
+#SBATCH --mem=200G
+##SBATCH --partition=high-capacity
+##SBATCH --qos=hc-concurrent-jobs
 
 #sphae install
-#sphae run --input tests/data/illumina-subset --threads 64 -k --use-conda --conda-frontend mamba 
-#sphae run --input tests/data/nanopore-subset --sequencing longread --threads 64 -k --conda-frontend mamba 
-#sphae run --input tests/data/nanopore-subset --sequencing longread --threads 64 -k --no_medaka --use-conda 
-sphae annotate --genome tests/data/genome --threads 64 -k --use-conda --conda-frontend mamba
+export VVDB=sphae/workflow/databases/Pfam35.0/Pfam-A.hmm.gz
+export CHECKVDB=sphae/workflow/databases/checkv-db-v1.5
+export PHAROKKADB=sphae/workflow/databases/pharokka_db
+export PHYNTENYDB=sphae/workflow/databases/phynteny_models_zenodo
+export PHOLDDB=sphae/workflow/databases/phold
 
-#sphae run --input tests/data/illumina-subset/ --threads 64 -k --db_dir tests/db --conda-create-envs-only
+#sphae run --input tests/data/illumina-subset --threads 32 -k --use-conda --conda-frontend mamba 
+#sphae run --input tests/data/nanopore-subset --sequencing longread --threads 64 -k --use-conda --conda-frontend mamba 
+sphae run --input tests/data/nanopore-subset --sequencing longread --threads 64 -k --no_medaka --use-conda --conda-frontend mamba 
+sphae annotate --genome tests/data/genome --threads 64 -k --use-conda --conda-frontend mamba