+++ sphae/__main__.py
@@ -107,11 +107,9 @@ ANNOTATE EXAMPLES
 \b
 sphae anntoate --genome <genomes>  
 sphae annotate --genome <genomes> --output <output> #define output directory
-sphae annotate --genome <genomes> --output <output> --db <database> #define database path
 """
 
 @click.command(epilog=help_msg_install, context_settings=dict(help_option_names=["-h", "--help"], ignore_unknown_options=True))
-@click.option('--db_dir', 'db_dir', help='Custom database directory', type=click.Path(), required=False)
 @click.option('--output', 'output', help='Output directory', type=click.Path(), default='sphae.out', show_default=True)
 @click.option("--configfile", default="config.yaml", show_default=False, callback=default_to_output,help="Custom config file [default: (outputDir)/config.yaml]",)
 @click.option('--threads', help='Number of threads to use', default=1, show_default=True)